rename styled component file and update imports for manual connection form
