+++ redisinsight/ui/src/pages/home/components/ManualConnection.styles.ts
