+++ redisinsight/ui/src/pages/home/components/manual-connection/manual-connection-form/ManualConnectionForm.tsx
@@ -33,7 +33,7 @@ import CloneConnection from './components/CloneConnection'
 import FooterActions from './components/FooterActions'
 import { AddConnection, EditConnection, EditSentinelConnection } from './forms'
 
-import { ScrollableWrapper, ContentWrapper } from '../../styles.module'
+import { ScrollableWrapper, ContentWrapper } from '../../ManualConnection.styles'
 
 export interface Props {
   formFields: DbConnectionInfo
